refactor(fhir): TEFCA QHIN Phase B — split tefca-ias into a registry-driven dispatcher
Phase B-1: pure structural refactor of the 2802-line edge function into
six focused modules. Same wire behavior, same audit + consent semantics,
but the giant per-resource if/else chain is gone.

Module split (supabase/functions/tefca-ias/)
- shared.ts     163 LOC  corsHeaders, fhir json helpers, SYSTEM_IDENTIFIERS,
                         US_CORE, ExchangePurpose, TEFCAContext, createBundle,
                         createOperationOutcome, fhirJsonResponse, errorResponse
- audit.ts       59 LOC  logTEFCAAccess + verifyPatientConsent
- mappers.ts    941 LOC  14 row -> FHIR R4 mappers (US Core 7.0 profiled),
                         pure functions, no I/O
- registry.ts   290 LOC  declarative ResourceConfig per FHIR resource type
                         (table, orderColumn, searchParams, profile, mapper,
                         interactions, customHandler flag)
- capability.ts  52 LOC  CapabilityStatement generated by walking the
                         registry — no second source of truth
- index.ts      877 LOC  slim router with three custom inline handlers
                         (Patient read/search/$everything, Observation
                         vitals+SDOH union, DiagnosticReport lab_orders +
                         lab_results join) and a registry-driven generic
                         handler that drives the remaining 11 resource types

Net: index.ts shrinks from 2802 -> 877 LOC (-2593 / +668). Total Deno code
across 6 files is 2382 LOC, ~420 LOC less than the previous megafile.

Behavior changes: none at the wire. Adding a new resource type now means
adding one entry to RESOURCE_REGISTRY and (if it doesn't fit the
patient_id-keyed pattern) a customHandler flag. CapabilityStatement and
dispatch update automatically.

Latent fix folded in: mapMedicationRequestToFHIR now falls back to
dispense.item_name when dispense.drug is undefined (the dispenses table
column is item_name; the previous mapper rendered an empty
medicationCodeableConcept.text in production).

Deferred to Phase B-2 (next PR):
- _history / vread interactions + resource_versions snapshot table
- Patient/$match (port from src/services/fhir/tefcaAuth.ts)
- Per-resource handler files (resources/*.ts) — current inline custom
  handlers for Patient/Observation/DiagnosticReport stay in index.ts

Verification
- npx tsc --noEmit clean
- npx eslint --max-warnings=0 supabase/functions/tefca-ias/ clean
- src/services tests still pass (allergies, labs)
- Manual review: Patient/$everything fan-out, all search routes,
  consent gate, audit log writes preserved
